polymer/requirements.txt
dj-database-url==0.5.0
Django==2.1
django-cors-headers==2.4.0
django-filter==2.0.0
django-rest-auth==0.9.3
django-storages==1.6.6
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
psycopg2==2.7.5
PyJWT==1.6.4
pytz==2018.5
six==1.11.0
